Add files via upload
I got a little confused on this homework and didn't realize until so close to the due date. I'll be in office hours next week to determine what went wrong and will look to use the resources better so this doesn't become a habit. Also, I am a little unsure of my GitHub usage so I really hope this is in the right place.